requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.6.3
asgiref==3.9.1
async-timeout==4.0.3
asyncpg==0.31.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
Babel==2.9.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
cachetools==4.2.2
certifi==2025.8.3
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
cloudinary==1.44.1
constantly==23.10.4
contourpy==1.3.2
crispy-bootstrap5==2026.3
cryptography==46.0.5
cycler==0.12.1
daphne==4.2.1
Django==5.2.6
django-cors-headers==4.9.0
django-crispy-forms==2.6
django-filter==25.2
djangorestframework==3.16.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.121.0
Flask==3.1.3
fonttools==4.60.0
frozenlist==1.7.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
magic-filter==1.0.12
MarkupSafe==3.0.3
matplotlib==3.10.6
msgpack==1.1.2
multidict==6.6.4
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
propcache==0.3.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.5.3
pydantic_core==2.14.6
pyodbc==5.3.0
pyOpenSSL==26.0.0
pyparsing==3.2.4
pyTelegramBotAPI==4.29.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
python-telegram-bot==22.5
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.3
redis==7.4.0
requests==2.32.5
rsa==4.9.1
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.15.3
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sqlparse==0.5.3
starlette==0.49.3
threadpoolctl==3.6.0
tomli==2.4.1
tornado==6.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==4.3.1
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.7
yarl==1.20.1
zope.interface==8.2
